refactor: Move ImageLinkConverter class to segment export namespace. Move SegmentExporterFileNameGenerator class to segment export namespace.
