+++ demosplan/DemosPlanCoreBundle/Logic/Segment/Export/SegmentExporterFileNameGenerator.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * All rights reserved
  */
 
-namespace demosplan\DemosPlanCoreBundle\Logic\Segment;
+namespace demosplan\DemosPlanCoreBundle\Logic\Segment\Export;
 
 use Cocur\Slugify\Slugify;
 use DemosEurope\DemosplanAddon\Contracts\Entities\UserInterface;
